# docs/swagger.yaml
basePath: /
definitions:
...
  blog.Blog:
    properties:
      authorId:
        type: integer
      categoryId:
        type: integer
      content:
        type: string
      createdAt:
        type: string
      excerpt:
        type: string
      id:
        type: integer
      location:
        type: string
      metadata:
        $ref: '#/definitions/blog.BlogMetadata'
      publishDate:
        type: string
      reviewedAt:
        type: string
      reviewedBy:
        type: integer
      slug:
        type: string
      status:
        $ref: '#/definitions/blog.BlogStatus'
      tags:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
  blog.BlogListResponse:
    properties:
      blogs:
        items:
          $ref: '#/definitions/blog.Blog'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
  blog.BlogMetadata:
    properties:
      keywords:
        items:
          type: string
        type: array
      metaDescription:
        type: string
      metaTitle:
        type: string
    type: object
  blog.BlogResponse:
    properties:
      blog:
        $ref: '#/definitions/blog.Blog'
    type: object
  blog.BlogStatus:
    enum:
    - draft
    - review
    - published
    type: string
    x-enum-varnames:
    - StatusDraft
    - StatusReview
    - StatusPublished
  blog.CreateBlogRequest:
    properties:
      authorId:
        type: integer
      categoryId:
        type: integer
      content:
        minLength: 100
        type: string
      excerpt:
        maxLength: 500
        minLength: 50
        type: string
      location:
        type: string
      metadata:
        $ref: '#/definitions/blog.BlogMetadata'
      publishDate:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/blog.BlogStatus'
        enum:
        - draft
        - review
        - published
      tags:
        type: string
      title:
        maxLength: 200
        minLength: 10
        type: string
    required:
    - authorId
    - categoryId
    - content
    - excerpt
    - publishDate
    - status
    - title
    type: object
  blog.UpdateBlogRequest:
    properties:
      authorId:
        type: integer
      categoryId:
        type: integer
      content:
        minLength: 100
        type: string
      excerpt:
        maxLength: 500
        minLength: 50
        type: string
      location:
        type: string
      metadata:
        $ref: '#/definitions/blog.BlogMetadata'
      publishDate:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/blog.BlogStatus'
        enum:
        - draft
        - review
        - published
      tags:
        type: string
      title:
        maxLength: 200
        minLength: 10
        type: string
    type: object
...
  response.Meta:
    properties:
      cursor:
        type: string
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
...
host: 192.168.1.2:8081
info:
  contact:
    email: support@example.com
    name: API Support
  description: RESTful API for healthcare market research reports and categories with
    caching and pagination support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://example.com/terms/
  title: Healthcare Market Research API
  version: "1.0"
paths:
...
  /api/v1/blogs:
    get:
      consumes:
      - application/json
      description: Get a paginated list of blogs with optional filtering
      parameters:
      - description: 'Filter by status: draft, review, published'
        in: query
        name: status
        type: string
      - description: Filter by category ID
        in: query
        name: categoryId
        type: integer
      - description: Filter by tags (comma-separated)
        in: query
        name: tags
        type: string
      - description: Filter by author ID
        in: query
        name: authorId
        type: integer
      - description: Filter by location
        in: query
        name: location
        type: string
      - description: Search in title, excerpt, content
        in: query
        name: search
        type: string
      - description: 'Page number (default: 1, min: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of blogs with pagination
          schema:
            $ref: '#/definitions/blog.BlogListResponse'
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      summary: Get all blogs
      tags:
      - Blogs
    post:
      consumes:
      - application/json
      description: Create a new blog post
      parameters:
      - description: Blog data
        in: body
        name: blog
        required: true
        schema:
          $ref: '#/definitions/blog.CreateBlogRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Blog created successfully
          schema:
            $ref: '#/definitions/blog.BlogResponse'
        "400":
          description: Bad request - invalid input or validation error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      summary: Create blog
      tags:
      - Blogs
  /api/v1/blogs/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a blog post by ID
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Blog deleted successfully
        "400":
          description: Bad request - invalid ID
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Blog not found
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      summary: Delete blog
      tags:
      - Blogs
    get:
      consumes:
      - application/json
      description: Get a single blog post by ID
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Blog details
          schema:
            $ref: '#/definitions/blog.BlogResponse'
        "400":
          description: Bad request - invalid ID
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Blog not found
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      summary: Get blog by ID
      tags:
      - Blogs
    put:
      consumes:
      - application/json
      description: Update an existing blog post
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: integer
      - description: Blog update data
        in: body
        name: blog
        required: true
        schema:
          $ref: '#/definitions/blog.UpdateBlogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Blog updated successfully
          schema:
            $ref: '#/definitions/blog.BlogResponse'
        "400":
          description: Bad request - invalid input or validation error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Blog not found
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      summary: Update blog
      tags:
      - Blogs
  /api/v1/blogs/{id}/publish:
    patch:
      consumes:
      - application/json
      description: Change blog status to published
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Blog published successfully
          schema:
            $ref: '#/definitions/blog.BlogResponse'
        "400":
          description: Bad request - invalid ID or blog cannot be published
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Blog not found
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      summary: Publish blog
      tags:
      - Blogs
  /api/v1/blogs/{id}/submit-review:
    patch:
      consumes:
      - application/json
      description: Change blog status from draft to review
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Blog submitted for review successfully
          schema:
            $ref: '#/definitions/blog.BlogResponse'
        "400":
          description: Bad request - invalid ID or blog cannot be submitted
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Blog not found
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      summary: Submit blog for review
      tags:
      - Blogs
  /api/v1/blogs/{id}/unpublish:
    patch:
      consumes:
      - application/json
      description: Change blog status from published back to draft
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Blog unpublished successfully
          schema:
            $ref: '#/definitions/blog.BlogResponse'
        "400":
          description: Bad request - invalid ID or blog is not published
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Blog not found
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                error:
                  type: string
              type: object
      summary: Unpublish blog
      tags:
      - Blogs
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...